+++ openaerostruct/utils/plot_wing.py
@@ -1,4 +1,4 @@
-""" Script to plot results from aero, struct, or aerostruct optimization.
+"""Script to plot results from aero, struct, or aerostruct optimization.
 
 Usage is `plot_wing __name__` for user-named database.
 
@@ -9,7 +9,6 @@ The larger the number, the closer the view. Floats or ints are accepted.
 Ex: `plot_wing aero.db 1` a wider view than `plot_wing aero.db 5`.
 
 """
-
 
 import sys
 import numpy as np
@@ -257,6 +256,9 @@ class Display(object):
             self.symmetry = True
         else:
             self.symmetry = False
+
+        # Temp Fix for FFD meshes
+        self.symmetry = True
 
         if self.symmetry:
             new_mesh = []
